Bug fixes.
State management with rxjs -> Home component takes from the route the teamId, and updates a teamId state, implemented as a BehaviourSubject observable. The app can subscribe to it, and use it anywhere. Now, teams drowpdown subscribes to it, and selects the relevant option. When attempting to a url with a teamId on it, the dropdown now reflects that.

Highlight rendering fix. Component was rendering twice, first without data being received, and change detection rerendered everything again with data received. So, Now component doesn't show anything until it gets valid data.